Refactoring AzazelRobotConfiguration to match other configs
