+++ TeamCode/src/main/java/intothedeep/teleop/AzazelTeleOp.java
@@ -1,8 +1,7 @@
 package intothedeep.teleop;
 
-import intothedeep.IntoTheDeepRobotConfiguration;
+import intothedeep.AzazelRobotConfiguration;
 import intothedeep.Constants;
-import intothedeep.teleop.AzazelRobotCapabilities;
 import t10.bootstrap.TeleOpOpMode;
 import t10.gamepad.GController;
 import t10.novel.mecanum.MecanumDriver;
@@ -10,8 +9,6 @@ import t10.novel.mecanum.MecanumDriver;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
-import org.firstinspires.ftc.robotcore.external.navigation.AxesOrder;
-import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
 import org.firstinspires.ftc.robotcore.external.navigation.YawPitchRollAngles;
 
 @TeleOp(name = "Azazel Tele-Op")
